+++ app/controllers/posts_controller.rb
@@ -1,5 +1,4 @@
 class PostsController < ApplicationController
-  
   def index
   end
