--- app/helpers/cards_helper.rb
@@ -1,2 +1,0 @@
-module CardsHelper
-end
