+++ app/models/card.rb
@@ -1,5 +1,2 @@
 class Card < ApplicationRecord
-
-  belongs_to :user, optional: true
-
 end
