+++ app/models/user.rb
@@ -6,6 +6,5 @@ class User < ApplicationRecord
 
   has_one :address
   has_one :profile
-  has_many :cards
 
 end
