--- app/views/cards/create.html.haml
@@ -1,1 +1,0 @@
-完了です。
