+++ app/views/signup/step3.html.haml
@@ -3,7 +3,7 @@
   .contents
     .content-header
       %h1 発送元・お届け先住所入力
-    = form_with model: @address, url: new_card_path, method: :get, local: true do |f|
+    = form_with model: @address, url: step4_signup_index_path, local: true do |f|
       .content-form
         %label.text お名前
         %span.required 必須
