+++ config/routes.rb
@@ -12,7 +12,4 @@ Rails.application.routes.draw do
       post "step4"
     end
   end
-
-  resources :cards, only: [:new, :create]
-
 end
