+++ db/migrate/20191216030523_create_cards.rb
@@ -1,11 +1,10 @@
 class CreateCards < ActiveRecord::Migration[5.2]
   def change
     create_table :cards do |t|
-      t.references :user, foreign_key: true
+      t.integer :user_id
       t.string :customer_id
       t.string :card_id
       t.timestamps
     end
   end
 end
-
